+++ src/61date.js
@@ -5,22 +5,28 @@
 //
 
 
-stdfn.NOW = function(){return d.getFullYear();};
+stdfn.NOW = function(){
+	var d = new Date();
+	var s = d.getFullYear()+"."+("0"+d.getMonth()).substr(-2)+"."+("0"+d.getDate()).substr(-2);
+	s += " "+("0"+d.getHours()).substr(-2)+":"+("0"+d.getMinutes()).substr(-2)+"."+("0"+d.getSeconds()).substr(-2);
+	return s;
+};
+
 
 stdfn.SECOND = function(d){
 	d = new Date(d);
-	return d.getSecond();
+	return d.getSeconds();
 };
 
 
 stdfn.MINUTE = function(d){
 	d = new Date(d);
-	return d.getMinute();
+	return d.getMinutes();
 };
 
 stdfn.HOUR = function(d){
 	d = new Date(d);
-	return d.getHour();
+	return d.getHours();
 };
 
 stdfn.DAY = stdfn.DAYOFMONTH = function(d){