+++ test/test133.js
@@ -16,19 +16,19 @@ describe('Test 133 SELECT VALUE, ROW, COLUMN, MATRIX', function() {
 
 	it("2. SELECT", function(done) {
 		var res = alasql('SELECT * FROM one');
-		assert.deepEqual(res1,[{a:1,b:"One"},{a:2,b:"Two"},{a:3,b:"Three"}]);
+		assert.deepEqual(res,[{a:1,b:"One"},{a:2,b:"Two"},{a:3,b:"Three"}]);
 
 		var res = alasql('SELECT VALUE * FROM one');
-		assert.deepEqual(res1,1);
+		assert.deepEqual(res,1);
 
 		var res = alasql('SELECT ROW * FROM one');
-		assert.deepEqual(res1,[1,"One"]);
+		assert.deepEqual(res,[1,"One"]);
 
 		var res = alasql('SELECT COLUMN * FROM one');
-		assert.deepEqual(res1,[1,2,3]);
+		assert.deepEqual(res,[1,2,3]);
 
 		var res = alasql('SELECT MATRIX * FROM one');
-		assert.deepEqual(res1,[[1,"One"],[2,"Two"],[3,"Three"]]);
+		assert.deepEqual(res,[[1,"One"],[2,"Two"],[3,"Three"]]);
 
 		done();
 	});